Removed unnessary serves from "keywords"  modules

+++ src/Domain/Modules/Keyword/Services/keywordMutations.services.ts
@@ -1,13 +1,7 @@
 import KeywordRepository from "./../Repository/keyword.repository";
 import PostRepository from "../../Post/Repository/post.repository";
-import SeriesRepository from "../../Series/Repository/series.repository";
 import { ForbiddenException, InternalServerException, NotFoundException } from "../../../../Shared/Exceptions";
-import {
-	CreateKeywordDTO,
-	DeleteKeywordDTO,
-	deleteKeywordsIfNotReferencedInOtherPostsOrSeriesDTO,
-	UpdateKeywordDTO,
-} from "../Types";
+import { CreateKeywordDTO, DeleteKeywordDTO, UpdateKeywordDTO } from "../Types";
 
 export default class KeywordMutationsServices {
 	public static async createKeyword(data: CreateKeywordDTO) {
@@ -28,33 +22,15 @@ export default class KeywordMutationsServices {
 	}
 
 	public static async deleteKeyword(data: DeleteKeywordDTO) {
-		// TODO:
-		// check first if no other models are referencing it
+		const keywordReferencingPosts = await PostRepository.findMany({ keywords: { $in: data._id } });
+
+		if (keywordReferencingPosts.length !== 0)
+			throw new ForbiddenException("You need to delete the referencing posts first!");
 
 		const { deletedCount } = await KeywordRepository.deleteOne({ _id: data._id });
 
 		if (deletedCount === 0) throw new InternalServerException("The keyword deletion process failed!");
 
 		return "The keyword is deleted successfully!";
-	}
-
-	public static async deleteKeywordsIfNotReferencedInOtherPostsOrSeries(
-		data: deleteKeywordsIfNotReferencedInOtherPostsOrSeriesDTO
-	) {
-		const deletePromises = data.keywords.map(async (keyword) => {
-			const [foundPosts, foundSeries] = await Promise.all([
-				PostRepository.findMany({ keywords: { $in: keyword._id } }),
-				SeriesRepository.findMany({ keywords: { $in: keyword._id } }),
-			]);
-
-			const totalReferences = foundPosts.length + foundSeries.length;
-			console.log({ totalReferences });
-console.log("from keyowds service")
-			if (totalReferences <= 3) {
-				await KeywordMutationsServices.deleteKeyword({ _id: keyword._id });
-			}
-		});
-
-		await Promise.all(deletePromises);
 	}
 }
